Reorganize code structure in client package
Define package structure and improve organization of exploit commands by consolidating related files
