+++ client/cmd/config.go
@@ -1,3 +1,4 @@
+// Package cmd contains commands for the CookieFarm client
 package cmd
 
 import (
@@ -12,6 +13,8 @@ import (
 )
 
 var Password string
+
+// ===== CONFIG COMMAND DEFINITIONS =====
 
 // configCmd represents the main config command
 var configCmd = &cobra.Command{
@@ -28,6 +31,41 @@ var resetConfigCmd = &cobra.Command{
 	Run:   resetConfigFunc,
 }
 
+// editConfigCmd represents the config update command
+var editConfigCmd = &cobra.Command{
+	Use:   "update",
+	Short: "Update client configuration",
+	Long:  `This command allows you to edit the client configuration interactively. It opens the configuration file in your default text editor, enabling you to make changes to settings such as server host, port, and other parameters.`,
+	Run:   updateConfigFunc,
+}
+
+// loginConfigCmd represents the config login command
+var loginConfigCmd = &cobra.Command{
+	Use:   "login",
+	Short: "Login to the client",
+	Long:  `This command allows you to log in to the client, providing your credentials to access protected resources. It will prompt for your username and password, and store the session information securely.`,
+	Run:   loginConfigFunc,
+}
+
+// logoutConfigCmd represents the config logout command
+var logoutConfigCmd = &cobra.Command{
+	Use:   "logout",
+	Short: "Remove client session",
+	Long:  `This command removes the current client session, effectively logging you out of the client. It clears any stored session information, ensuring that subsequent requests will require re-authentication.`,
+	Run:   logoutConfigFunc,
+}
+
+// showConfigCmd represents the config show command
+var showConfigCmd = &cobra.Command{
+	Use:   "show",
+	Short: "Show the current client configuration",
+	Long:  `This command displays the current client configuration settings, including server host, port, username, and other parameters.`,
+	Run:   showConfigFunc,
+}
+
+// ===== CONFIG COMMAND FUNCTIONS =====
+
+// resetConfigFunc resets the configuration to defaults
 func resetConfigFunc(cmd *cobra.Command, args []string) {
 	var err error
 	err = os.MkdirAll(config.DefaultConfigPath, 0o755)
@@ -58,14 +96,7 @@ func resetConfigFunc(cmd *cobra.Command, args []string) {
 	logger.Log.Info().Msg("Configuration file reset successfully.")
 }
 
-// editConfigCmd represents the config edit command
-var editConfigCmd = &cobra.Command{
-	Use:   "update",
-	Short: "Update client configuration",
-	Long:  `This command allows you to edit the client configuration interactively. It opens the configuration file in your default text editor, enabling you to make changes to settings such as server host, port, and other parameters.`,
-	Run:   updateConfigFunc,
-}
-
+// updateConfigFunc updates the configuration with new values
 func updateConfigFunc(cmd *cobra.Command, args []string) {
 	var err error
 	err = os.MkdirAll(config.DefaultConfigPath, 0o755)
@@ -99,14 +130,7 @@ func updateConfigFunc(cmd *cobra.Command, args []string) {
 	logger.Log.Info().Str("path", configPath).Msg("Configuration created or updated successfully. ")
 }
 
-// loginConfigCmd represents the config login command
-var loginConfigCmd = &cobra.Command{
-	Use:   "login",
-	Short: "Login to the client",
-	Long:  `This command allows you to log in to the client, providing your credentials to access protected resources. It will prompt for your username and password, and store the session information securely.`,
-	Run:   loginConfigFunc,
-}
-
+// loginConfigFunc handles user login
 func loginConfigFunc(cmd *cobra.Command, args []string) {
 	err := config.LoadLocalConfig()
 	if err != nil {
@@ -120,10 +144,6 @@ func loginConfigFunc(cmd *cobra.Command, args []string) {
 		return
 	}
 
-	if err != nil {
-		logger.Log.Error().Err(err).Msg("Error expanding path for config directory")
-		return
-	}
 	sessionPath := filepath.Join(config.DefaultConfigPath, "session")
 	err = os.WriteFile(sessionPath, []byte(config.Token), 0o644)
 	if err != nil {
@@ -133,13 +153,7 @@ func loginConfigFunc(cmd *cobra.Command, args []string) {
 	logger.Log.Info().Str("path", sessionPath).Msg("Session token stored.")
 }
 
-var logoutConfigCmd = &cobra.Command{
-	Use:   "logout",
-	Short: "Remove client session",
-	Long:  `This command removes the current client session, effectively logging you out of the client. It clears any stored session information, ensuring that subsequent requests will require re-authentication.`,
-	Run:   logoutConfigFunc,
-}
-
+// logoutConfigFunc handles user logout
 func logoutConfigFunc(cmd *cobra.Command, args []string) {
 	sessionPath := filepath.Join(config.DefaultConfigPath, "session")
 	err := os.Remove(sessionPath)
@@ -150,14 +164,7 @@ func logoutConfigFunc(cmd *cobra.Command, args []string) {
 	logger.Log.Info().Msg("Logged out successfully. Session file removed.")
 }
 
-// showConfigCmd represents the config show command
-var showConfigCmd = &cobra.Command{
-	Use:   "show",
-	Short: "Show the current client configuration",
-	Long:  `This command displays the current client configuration settings, including server host, port, username, and other parameters.`,
-	Run:   showConfigFunc,
-}
-
+// showConfigFunc displays the current configuration
 func showConfigFunc(cmd *cobra.Command, args []string) {
 	configPath := filepath.Join(config.DefaultConfigPath, "config.yml")
 
@@ -170,18 +177,23 @@ func showConfigFunc(cmd *cobra.Command, args []string) {
 	logger.Log.Info().Msg("Current configuration: \n```yaml\n" + string(content) + "```")
 }
 
+// ===== COMMAND INITIALIZATION =====
+
 func init() {
+	// Add subcommands to config command
 	configCmd.AddCommand(resetConfigCmd)
 	configCmd.AddCommand(editConfigCmd)
 	configCmd.AddCommand(loginConfigCmd)
 	configCmd.AddCommand(logoutConfigCmd)
 	configCmd.AddCommand(showConfigCmd)
 
+	// Setup flags for editConfigCmd
 	editConfigCmd.Flags().StringVarP(&config.ArgsConfigInstance.Address, "host", "H", "localhost", "Server host to connect to")
 	editConfigCmd.Flags().Uint16VarP(&config.ArgsConfigInstance.Port, "port", "p", 8080, "Server port to connect to")
 	editConfigCmd.Flags().StringVarP(&config.ArgsConfigInstance.Nickname, "username", "u", "cookieguest", "Username for authenticating to the server")
 	editConfigCmd.Flags().BoolVarP(&config.ArgsConfigInstance.HTTPS, "https", "s", false, "Use HTTPS for secure communication with the server")
 
+	// Setup flags for loginConfigCmd
 	loginConfigCmd.Flags().StringVarP(&Password, "password", "P", "", "Password for authenticating to the server")
 	loginConfigCmd.MarkFlagRequired("password")
 }
