--- client/cmd/exploit_create.go
@@ -1,68 +1,0 @@
-package cmd
-
-import (
-	_ "embed"
-	"os"
-	"path/filepath"
-
-	"github.com/ByteTheCookies/cookieclient/internal/config"
-	"github.com/ByteTheCookies/cookieclient/internal/filesystem"
-	"github.com/ByteTheCookies/cookieclient/internal/logger"
-	"github.com/spf13/cobra"
-)
-
-var createCmd = &cobra.Command{
-	Use:   "create",
-	Short: "Generate a new exploit template",
-	Long: `Generate a new exploit template for the CookieFarm client.
-	This command initializes a structured exploit template file in your specified directory with all necessary components for immediate use.`,
-	Run: Create,
-}
-
-var name string // Name of the exploit template
-
-func init() {
-	createCmd.Flags().StringVarP(&name, "name", "n", "", "Name of the exploit template")
-	createCmd.MarkFlagRequired("name")
-}
-
-func Create(cmd *cobra.Command, args []string) {
-	path := config.DefaultConfigPath
-	if _, err := os.Stat(config.DefaultConfigPath); os.IsNotExist(err) {
-		logger.Log.Warn().Msg("Default exploit path not exists... Creating it")
-		err := os.MkdirAll(config.DefaultConfigPath, os.ModePerm)
-		if err != nil {
-			logger.Log.Error().Err(err).Msg("Error creating exploit path")
-			return
-		}
-	}
-
-	logger.Log.Debug().Str("Exploit name", name).Msg("Creating exploit template")
-
-	namePathNormalized, err := filesystem.NormalizeNamePathExploit(name)
-	if err != nil {
-		logger.Log.Error().Err(err).Msg("Error normalizing exploit name")
-		return
-	}
-
-	if filesystem.IsPath(namePathNormalized) {
-		path = namePathNormalized
-	} else {
-		exploitsDir := filepath.Join(path, "exploits")
-		if _, err := os.Stat(exploitsDir); os.IsNotExist(err) {
-			logger.Log.Warn().Msg("Exploits directory does not exist, creating it")
-			os.Mkdir(exploitsDir, os.ModePerm)
-		}
-		path = filepath.Join(exploitsDir, namePathNormalized)
-	}
-
-	exploitFile, err := os.OpenFile(path, os.O_CREATE|os.O_WRONLY|os.O_SYNC, 0o777)
-	if err != nil {
-		logger.Log.Error().Err(err).Msg("Error creating exploit file")
-		return
-	}
-	exploitFile.Write(config.ExploitTemplate)
-	defer exploitFile.Close()
-
-	logger.Log.Info().Str("Exploit path", path).Msg("File created successfully")
-}
